ai-service/requirements.txt
# MongoDB Log Monitor - AI Service Dependencies

# Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0

# MongoDB
pymongo==4.6.1

# Kafka
kafka-python==2.0.2

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.2

# NLP
spacy==3.7.2

# Data Processing
python-dateutil==2.8.2
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2

# Logging
colorlog==6.8.0

# Download spaCy model
# python -m spacy download en_core_web_sm

